# rework MG DoF distribution