--- a/tests/multigrid/constrained_dofs_02.mpirun=2.output
+++ b/tests/multigrid/constrained_dofs_02.mpirun=2.output
@@ -1,0 +1,125 @@
+
+DEAL:0::FE_Q<2>(1)
+DEAL:0::Level 0:
+DEAL:0::get_refinement_edge_indices:
+DEAL:0::{}
+DEAL:0::get_boundary_indices:
+DEAL:0::{[0,3]}
+DEAL:0::relevant:
+DEAL:0::{[0,3]}
+DEAL:0::active:
+DEAL:0::{[0,3]}
+DEAL:0::ok ok 
+DEAL:0::Level 1:
+DEAL:0::get_refinement_edge_indices:
+DEAL:0::{}
+DEAL:0::get_boundary_indices:
+DEAL:0::{[0,2], [4,8]}
+DEAL:0::relevant:
+DEAL:0::{[0,8]}
+DEAL:0::active:
+DEAL:0::{[0,3]}
+DEAL:0::ok ok 
+DEAL:0::Level 2:
+DEAL:0::get_refinement_edge_indices:
+DEAL:0::{}
+DEAL:0::get_boundary_indices:
+DEAL:0::{[0,2], 4, 6, 9, 15}
+DEAL:0::relevant:
+DEAL:0::{[0,10], 13, [15,17]}
+DEAL:0::active:
+DEAL:0::{[0,7]}
+DEAL:0::ok ok 
+DEAL:0::Level 3:
+DEAL:0::get_refinement_edge_indices:
+DEAL:0::{}
+DEAL:0::get_boundary_indices:
+DEAL:0::{[0,2], 4, 6, 9, 11, 15, 18, 25}
+DEAL:0::relevant:
+DEAL:0::{[0,23], [25,26], 29, 31}
+DEAL:0::active:
+DEAL:0::{[0,17]}
+DEAL:0::ok ok 
+DEAL:0::Level 4:
+DEAL:0::get_refinement_edge_indices:
+DEAL:0::{0, 2, 6, [14,16], 19, [29,31], [35,36], 38, [40,42], 54}
+DEAL:0::get_boundary_indices:
+DEAL:0::{[0,1], 4, 15, 17, 21, 33, 35, 43}
+DEAL:0::relevant:
+DEAL:0::{[0,39], [43,45], 49, [53,54]}
+DEAL:0::active:
+DEAL:0::{[0,25], [29,42]}
+DEAL:0::ok ok 
+DEAL:0::Level 5:
+DEAL:0::get_refinement_edge_indices:
+DEAL:0::{0, 2, [4,6], [8,9], [11,12], [14,15], [17,23], [25,30], 33, [35,38], 40, [42,46], [48,50], 52, [54,55]}
+DEAL:0::get_boundary_indices:
+DEAL:0::{[0,1], 4, 29, 31, 35}
+DEAL:0::relevant:
+DEAL:0::{[0,50], [52,55]}
+DEAL:0::active:
+DEAL:0::{[0,43]}
+DEAL:0::ok ok 
+
+DEAL:1::FE_Q<2>(1)
+DEAL:1::Level 0:
+DEAL:1::get_refinement_edge_indices:
+DEAL:1::{}
+DEAL:1::get_boundary_indices:
+DEAL:1::{[0,3]}
+DEAL:1::relevant:
+DEAL:1::{[0,3]}
+DEAL:1::active:
+DEAL:1::{}
+DEAL:1::ok ok 
+DEAL:1::Level 1:
+DEAL:1::get_refinement_edge_indices:
+DEAL:1::{}
+DEAL:1::get_boundary_indices:
+DEAL:1::{[0,2], [4,8]}
+DEAL:1::relevant:
+DEAL:1::{[0,8]}
+DEAL:1::active:
+DEAL:1::{[1,8]}
+DEAL:1::ok ok 
+DEAL:1::Level 2:
+DEAL:1::get_refinement_edge_indices:
+DEAL:1::{}
+DEAL:1::get_boundary_indices:
+DEAL:1::{[0,2], 4, 6, 9, [11,12], [14,15], [18,20], [22,24]}
+DEAL:1::relevant:
+DEAL:1::{[0,24]}
+DEAL:1::active:
+DEAL:1::{[3,24]}
+DEAL:1::ok ok 
+DEAL:1::Level 3:
+DEAL:1::get_refinement_edge_indices:
+DEAL:1::{24, [27,28], 30, [32,34], [38,40], [42,44]}
+DEAL:1::get_boundary_indices:
+DEAL:1::{6, 9, 11, 15, 18, 25, 27, 35, 38}
+DEAL:1::relevant:
+DEAL:1::{3, [5,44]}
+DEAL:1::active:
+DEAL:1::{8, [11,44]}
+DEAL:1::ok ok 
+DEAL:1::Level 4:
+DEAL:1::get_refinement_edge_indices:
+DEAL:1::{14, 28, [35,36], 38, [40,42], [46,48], [50,52], [54,56]}
+DEAL:1::get_boundary_indices:
+DEAL:1::{17, 21, 33, 35, 43, 46}
+DEAL:1::relevant:
+DEAL:1::{[9,14], [16,25], [31,56]}
+DEAL:1::active:
+DEAL:1::{[12,14], [20,28], [43,56]}
+DEAL:1::ok ok 
+DEAL:1::Level 5:
+DEAL:1::get_refinement_edge_indices:
+DEAL:1::{23, [25,28], [35,38], 40, [42,46], [48,52], [54,56]}
+DEAL:1::get_boundary_indices:
+DEAL:1::{31, 35}
+DEAL:1::relevant:
+DEAL:1::{[23,28], [31,32], [34,56]}
+DEAL:1::active:
+DEAL:1::{26, [40,41], [43,56]}
+DEAL:1::ok ok 
+