Fix a rare case of missing level ghost neighbors for geometric
multigrid.

Add sanity checks to confirm several assumptions made in the code to
detect problems like this in the future.

Rework and simplify the logic of distribute_mg_dofs: don't communicate
ghost neighbors by level but all at once and instead of guessing who
will send data, actually request cells from the ghost neighbors.

Add tests.