@tjhei 2016-04-06
@kronbichler I think this fixes all the issues. I am not sure how the new algorithm compares to the old one, especially on a very large scale. Do you want to try things out before merging?