@kronbichler 2016-04-06
The patch looks good (apart from indentation). I'll test it on a larger scale.