@kronbichler 2016-04-06
@tjhei I tested the patch on my original problem (224 cores, 6840 cells on 7 levels in 3D) and it appears to deadlock in distribute_mg_dofs. While reducing the problem, I realized that it already appears on 28 cores. I'm trying to see if I can see it also in 2D and even fewer MPI ranks.