@kronbichler 2016-04-06
Given the memory problems that @bangerth discovered, I can imagine that this is my problem here since I'm using a different MPI than you (IBM MPI). I'll wait with further tests until those are fixed.