@tjhei 2016-04-06
Oops, that is what you get for trying to finish things in a hurry. Thanks for the comments. Updated.