@tjhei 2016-04-06
fixed warnings