@kronbichler 2016-04-06
It still deadlocks for 27 and 28 ranks (but not 8, 16, 17). I will investigate more tomorrow.