@tjhei 2016-04-06
Does it deadlock in the new debug code inside execute_c&r() or inside distribute_dofs()? If the problem is small enough to run on a single machine, shoot it my way. Thanks.