@kronbichler 2016-04-07
Here is a minimal code that reproduces the behavior (sorry for not uploading it directly to your branch, I'm testing this on my cluster without direct internet connection and have lost git when copying things over):
```
#include <deal.II/base/logstream.h>
#include <deal.II/base/timer.h>
#include <deal.II/fe/fe_q.h>
#include <deal.II/grid/tria_accessor.h>
#include <deal.II/grid/tria_iterator.h>
#include <deal.II/distributed/tria.h>
#include <deal.II/grid/grid_generator.h>
#include <deal.II/dofs/dof_handler.h>

using namespace dealii;

template <int dim>
void do_test ()
{
  for (unsigned int i=1; i<=Utilities::MPI::n_mpi_processes(MPI_COMM_WORLD); ++i)
    {
      MPI_Comm sub_comm;
      std::vector<int> ranks(i);
      for (unsigned int j=0; j<ranks.size(); ++j)
        ranks[j] = j;
      MPI_Group world_group;
      MPI_Comm_group(MPI_COMM_WORLD, &world_group);
      MPI_Group group;
      MPI_Group_incl(world_group, i, &ranks[0], &group);

      MPI_Comm_create_group(MPI_COMM_WORLD, group, 0, &sub_comm);

      if (Utilities::MPI::this_mpi_process(MPI_COMM_WORLD) >= i)
        continue;

      ConditionalOStream pcout(std::cout, Utilities::MPI::this_mpi_process(sub_comm) == 0);
      pcout << "Number of MPI processes: " << Utilities::MPI::n_mpi_processes(sub_comm)
            << std::endl;
      FE_Q<dim> fe(2);
      pcout << "Testing " << fe.get_name() << std::endl << std::endl;
      parallel::distributed::Triangulation<dim> triangulation
        (sub_comm,
         Triangulation<dim>::none,
         parallel::distributed::Triangulation<dim>::construct_multigrid_hierarchy);

      Timer time;
      GridGenerator::subdivided_hyper_cube (triangulation, 1, -1, 1);
      triangulation.refine_global(2);
      pcout << "Time grid generation uniform:  " << time.wall_time() << "s" << std::endl;
      time.restart();
      for (typename Triangulation<dim>::active_cell_iterator cell=triangulation.begin_active();
           cell != triangulation.end(); ++cell)
        if (cell->is_locally_owned() &&
            cell->center().norm() < 0.55)
          cell->set_refine_flag();
      triangulation.execute_coarsening_and_refinement();
      for (typename Triangulation<dim>::active_cell_iterator cell=triangulation.begin_active();
           cell != triangulation.end(); ++cell)
        if (cell->is_locally_owned() &&
            cell->center().norm() > 0.3 && cell->center().norm() < 0.42)
          cell->set_refine_flag();
      triangulation.execute_coarsening_and_refinement();
      for (typename Triangulation<dim>::active_cell_iterator cell=triangulation.begin_active();
          cell != triangulation.end(); ++cell)
        if (cell->is_locally_owned() &&
            cell->center().norm() > 0.335 && cell->center().norm() < 0.39)
          cell->set_refine_flag();
      triangulation.execute_coarsening_and_refinement();
      pcout << "Time grid generation adaptive: " << time.wall_time() << "s" << std::endl;

      time.restart();
      DoFHandler<dim> dof_handler(triangulation);
      dof_handler.distribute_dofs(fe);
      pcout << "Time distribute dofs: " << time.wall_time() << "s" << std::endl;

      time.restart();
      dof_handler.distribute_mg_dofs(fe);
      pcout << "Time distribute mgdofs: " << time.wall_time() << "s" << std::endl;
    }
}


int main (int argc, char** argv)
{
  Utilities::MPI::MPI_InitFinalize mpi(argc, argv, 1);
  do_test<3>();
  return 0;
}
```
This code runs fine up to 19 processors but on 20 processors it deadlocks in `distribute_mg_dofs`.