@tjhei 2016-04-12
Nasty problem, but I hope this fix will deal with related problems correctly. Thanks again for the testcase.