@kronbichler 2016-04-12
@tjhei I checked it on my standard problems in 2D/3D with double/float mg cycles and I can confirm it works now. I checked up to 3584 MPI ranks and 2b cells with 17b dofs right now but can't see anything go wrong if we would go further. (The old implementation was tested on up to 28*512=14k procs and 57b dofs where the bug did not appear.)

Regarding performance:
- On 224 cores and 272m dofs, the time for `distribute_mg_dofs` went from 4.03s to 1.97s (for comparison, distribute_dofs is 1.6s in this case and the grid creation from scratch 19s, both essentially unchanged from before).

- On 896 cores and 3.7b dofs, distribute_mg_dofs went from 14.0s to 6.69s.

- The total setup time (mesh generation, dof distribution, setup of vectors + matrix-free, assembly of rhs, setup of MG data structures) on 17b dofs/3584 procs takes 138 seconds, of which the meshing alone already takes 80s. Solving in 8 iterations takes 52s, a v-cycle is 6.06s.

Nice work!