@kronbichler 2016-04-12
From my point of view, this looks ready. Add a changelog entry and we are good to go.