@tjhei 2016-04-12
Nice, thanks for checking. I was not at all convinced this would improve performance, but 2x in some places is great news.

Squashed and changelog added.