@kronbichler 2016-04-12
I wait a bit with merging since I was partly involved in this. If nobody has additional comments within a day, I (or @tjhei) will merge.